+++ packages/experimental/src/components/ModifiedTabs/ModifiedTabs.stories.js
@@ -11,6 +11,9 @@ import { action } from '@storybook/addon-actions';
 
 import { ModifiedTabs } from '.';
 import { Modal, RadioButton, RadioButtonGroup } from 'carbon-components-react';
+import {
+  storybookPrefixCanary as storybookPrefix /* , storybookPrefixReleased */,
+} from '../../../config';
 
 import styles from './_storybook-styles.scss'; // import index in case more files are added later.
 
@@ -32,7 +35,7 @@ const commonStoryCode = {
 };
 
 export default {
-  title: 'Experimental/ModifiedTabs',
+  title: `${storybookPrefix}/ModifiedTabs`,
   component: ModifiedTabs,
   parameters: { styles },
 };