+++ packages/experimental/src/components/Tearsheet/TearsheetNarrow.stories.js
@@ -10,6 +10,7 @@ import React, { useState } from 'react';
 import { action } from '@storybook/addon-actions';
 
 import { pkgPrefix } from '../../global/js/settings';
+import { storybookPrefixCanary as storybookPrefix } from '../../../config';
 
 import { Button } from 'carbon-components-react';
 
@@ -21,7 +22,7 @@ import styles from './_storybook-styles.scss';
 import mdx from './Tearsheet.mdx';
 
 export default {
-  title: 'Experimental/Tearsheets/TearsheetNarrow',
+  title: `${storybookPrefix}/Tearsheets/TearsheetNarrow`,
   component: TearsheetNarrow,
   subcomponents: {
     Tearsheet,